+++ client/src/hooks/useOrigin.ts
@@ -5,7 +5,7 @@ import { calculateCenterPosition } from "../routes/editor/utils";
 export function useOrigin() {
   const dndContextRef = useRef<HTMLDivElement>(null);
   const { setOriginPosition } = useEditorStore();
-
+  
   function handleCenterOrigin() {
     if (dndContextRef.current) {
       const { width, height } = dndContextRef.current.getBoundingClientRect();
